move things where they need to be
