+++ src/components/List/list.jsx
@@ -1,0 +1,17 @@
+import * as React from 'react';
+import ListItem from '../ListItem/list-item';
+
+const List = ({ items }) => {
+  return (
+      <div className="nearbyCitiesList">
+        {items.map((item) => {
+          return <ListItem
+            cityItem={item}
+            key={item.id}/>;
+        })}
+      </div>
+  );
+
+};
+
+export default List;
