--- src/components/list.jsx
@@ -1,15 +1,0 @@
-var React = require('react');
-import ListItem from './list-item';
-require('../../pcss/main.css');
-
-const List = ({items}) => {
-  return <div className="nearbyCitiesList">
-    {items.map((item) => {
-      return <ListItem
-        cityItem={item}
-        key={item.id}/>
-    })}
-  </div>
-}
-
-export default List;
