+++ src/main.css.d.ts
@@ -1,0 +1,3 @@
+export const move: string;
+export const container: string;
+export const fadein: string;
